+++ janitor/functions.py
@@ -244,6 +244,7 @@ def clean_names(
     strip_accents: bool = True,
     preserve_original_columns: bool = True,
     enforce_string: bool = True,
+    truncate_limit: int = None,
 ) -> pd.DataFrame:
     """
     Clean column names.
@@ -296,6 +297,8 @@ def clean_names(
     :param enforce_string: Whether or not to convert all column names
         to string type. Defaults to True, but can be turned off.
         Columns with >1 levels will not be converted by default.
+    :param truncate_limit: (optional) Truncates formatted column names to
+        the specified length. Default None does not truncate.
     :returns: A pandas DataFrame.
     """
     original_column_names = list(df.columns)
@@ -315,6 +318,8 @@ def clean_names(
 
     df = df.rename(columns=lambda x: re.sub("_+", "_", x))  # noqa: PD005
     df = _strip_underscores(df, strip_underscores)
+
+    df = df.rename(columns=lambda x: x[:truncate_limit])
 
     # Store the original column names, if enabled by user
     if preserve_original_columns:
@@ -4693,7 +4698,7 @@ def pivot_longer(
     index: Optional[Union[List, Tuple, str, Pattern]] = None,
     column_names: Optional[Union[List, Tuple, str, Pattern]] = None,
     names_sep: Optional[Union[str, Pattern]] = None,
-    names_pattern: Optional[Union[str, Pattern]] = None,
+    names_pattern: Optional[Union[List, Tuple, str, Pattern]] = None,
     names_to: Optional[Union[List, Tuple, str]] = None,
     values_to: Optional[str] = "value",
 ) -> pd.DataFrame:
@@ -4803,6 +4808,13 @@ def pivot_longer(
         2  treat1         2         5
         3  treat2         3         4
 
+    Let's break down the `.value` idea a bit. When `.value` is used,
+    `pivot_longer` creates a pairing. In the example above, we get a pairing
+    {"group":["treat1", "treat2"], ".value":["measure1", "measure2"]}. All
+    the values associated with `.value` become new column names, while those
+    not associated with `.value`(`treat1` and `treat2`) become values in a
+    new column `group`. `values_to` is overridden during this process.
+
     Example 4: We can also pivot from wide to long using regular expressions
 
     .. code-block:: python
@@ -4819,6 +4831,15 @@ def pivot_longer(
         0     1  10.0  0.1
         1     2  20.0  0.2
         2     3  30.0  0.3
+
+    The same idea of `.value` works here as well. Based on the capturing groups
+    in the regex in `names_pattern`, we have two pairings -->
+    {".value":["n", "pct"], "name":[1,2,3]}. Just like in the previous example,
+    the values associated with `.value` become new column names, while those
+    not associated with `.value` become values in the new column ``name``.
+
+    Note that there are no limits to the pairing; however, you can only have
+    one `.value` in ``names_to``.
 
     You can also take advantage of `janitor.patterns` function, which allows
     selection of columns via a regular expression; this can come in handy if
@@ -4885,15 +4906,14 @@ def pivot_longer(
 
     :param df: A pandas dataframe.
     :param index: Name(s) of columns to use as identifier variables.
-        Should be either a single column name, a list/tuple of
-        column names, or a or a `janitor.patterns` function. You can
-        also dynamically select column names by using a regular
-        expression with the `janitor.patterns` function.
+        Should be either a single column name, or a list/tuple of
+        column names. You can also dynamically select column names
+        by using a regular expression with the `janitor.patterns`
+        function.
     :param column_names: Name(s) of columns to unpivot. Should be either
-        a single column name, a list/tuple of column names, or a
-        `janitor.patterns` function. You can also dynamically select
-        column names by using a regular expression with the
-        `janitor.patterns` function.
+        a single column name, a list/tuple of column names. You can also
+        dynamically select column names by using a regular expression
+        with the `janitor.patterns` function.
     :param names_to: Name of new column as a string that will contain
         what were previously the column names in `column_names`.
         The default is `variable` if no value is provided. It can
@@ -4906,9 +4926,16 @@ def pivot_longer(
         `names_to` contains multiple values. It takes the same
         specification as pandas' `str.split` method, and can be a string
         or regular expression.
-    :param names_pattern: Determines how the column name is broken up. It takes
-        the same specification as pandas' `str.extractall` method, which is
-        a regular expression containing matching groups.
+    :param names_pattern: Determines how the column name is broken up.
+        It can be a regular expression containing matching groups
+        matching the same specification as pandas' `str.extractall` method,
+        or a list/tuple of regular expressions, which devolves to
+        ``numpy.select`` and pandas' ``str.contains``. For a list of
+        regular expressions, ``names_to`` must be a list/tuple and the
+        lengths of both arguments must match. The entries in both arguments
+        must also match positionally, i.e  if `regex1` in `names_pattern` is
+        the first item, it will be aligned to `new_column_name_1` in `names_to`
+        if `new_column_name_1` is the first item, and so on.
     :param values_to: Name of new column as a string that will contain what
         were previously the values of the columns in `column_names`.
     :returns: A pandas DataFrame that has been unpivoted from wide to long